refactor: move all projects service into their own specific folders
