--- api/drizzle/0002_fluffy_power_man.sql
@@ -1,1 +1,0 @@
-ALTER TABLE "projects" ALTER COLUMN "id" DROP DEFAULT;
