--- api/drizzle/meta/0002_snapshot.json
@@ -1,271 +1,0 @@
-{
-  "id": "229d8d3b-e1b9-4e3e-b80f-491c1bb4e23f",
-  "prevId": "2185a7ae-146e-4020-8c41-449e5d48432e",
-  "version": "7",
-  "dialect": "postgresql",
-  "tables": {
-    "public.communes": {
-      "name": "communes",
-      "schema": "",
-      "columns": {
-        "insee_code": {
-          "name": "insee_code",
-          "type": "text",
-          "primaryKey": true,
-          "notNull": true
-        }
-      },
-      "indexes": {},
-      "foreignKeys": {},
-      "compositePrimaryKeys": {},
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    },
-    "public.projects": {
-      "name": "projects",
-      "schema": "",
-      "columns": {
-        "id": {
-          "name": "id",
-          "type": "uuid",
-          "primaryKey": true,
-          "notNull": true
-        },
-        "created_at": {
-          "name": "created_at",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": true,
-          "default": "now()"
-        },
-        "updated_at": {
-          "name": "updated_at",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": true,
-          "default": "now()"
-        },
-        "nom": {
-          "name": "nom",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "description": {
-          "name": "description",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "code_siret": {
-          "name": "code_siret",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "porteur_referent_email": {
-          "name": "porteur_referent_email",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "porteur_referent_telephone": {
-          "name": "porteur_referent_telephone",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "porteur_referent_prenom": {
-          "name": "porteur_referent_prenom",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "porteur_referent_nom": {
-          "name": "porteur_referent_nom",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "porteur_referent_fonction": {
-          "name": "porteur_referent_fonction",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "budget": {
-          "name": "budget",
-          "type": "integer",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "forecasted_start_date": {
-          "name": "forecasted_start_date",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "status": {
-          "name": "status",
-          "type": "project_status",
-          "typeSchema": "public",
-          "primaryKey": false,
-          "notNull": true
-        }
-      },
-      "indexes": {},
-      "foreignKeys": {},
-      "compositePrimaryKeys": {},
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    },
-    "public.projects_to_communes": {
-      "name": "projects_to_communes",
-      "schema": "",
-      "columns": {
-        "project_id": {
-          "name": "project_id",
-          "type": "uuid",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "commune_id": {
-          "name": "commune_id",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        }
-      },
-      "indexes": {
-        "commune_project_idx": {
-          "name": "commune_project_idx",
-          "columns": [
-            {
-              "expression": "commune_id",
-              "isExpression": false,
-              "asc": true,
-              "nulls": "last"
-            },
-            {
-              "expression": "project_id",
-              "isExpression": false,
-              "asc": true,
-              "nulls": "last"
-            }
-          ],
-          "isUnique": false,
-          "concurrently": false,
-          "method": "btree",
-          "with": {}
-        }
-      },
-      "foreignKeys": {
-        "projects_to_communes_project_id_projects_id_fk": {
-          "name": "projects_to_communes_project_id_projects_id_fk",
-          "tableFrom": "projects_to_communes",
-          "tableTo": "projects",
-          "columnsFrom": ["project_id"],
-          "columnsTo": ["id"],
-          "onDelete": "no action",
-          "onUpdate": "no action"
-        },
-        "projects_to_communes_commune_id_communes_insee_code_fk": {
-          "name": "projects_to_communes_commune_id_communes_insee_code_fk",
-          "tableFrom": "projects_to_communes",
-          "tableTo": "communes",
-          "columnsFrom": ["commune_id"],
-          "columnsTo": ["insee_code"],
-          "onDelete": "no action",
-          "onUpdate": "no action"
-        }
-      },
-      "compositePrimaryKeys": {
-        "projects_to_communes_project_id_commune_id_pk": {
-          "name": "projects_to_communes_project_id_commune_id_pk",
-          "columns": ["project_id", "commune_id"]
-        }
-      },
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    },
-    "public.services": {
-      "name": "services",
-      "schema": "",
-      "columns": {
-        "id": {
-          "name": "id",
-          "type": "uuid",
-          "primaryKey": true,
-          "notNull": true,
-          "default": "gen_random_uuid()"
-        },
-        "createdAt": {
-          "name": "createdAt",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": false,
-          "default": "now()"
-        },
-        "name": {
-          "name": "name",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "description": {
-          "name": "description",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "logoUrl": {
-          "name": "logoUrl",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "url": {
-          "name": "url",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        }
-      },
-      "indexes": {},
-      "foreignKeys": {},
-      "compositePrimaryKeys": {},
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    }
-  },
-  "enums": {
-    "public.permission_type": {
-      "name": "permission_type",
-      "schema": "public",
-      "values": ["EDIT", "VIEW"]
-    },
-    "public.project_status": {
-      "name": "project_status",
-      "schema": "public",
-      "values": ["IDEE", "FAISABILITE", "EN_COURS", "IMPACTE", "ABANDONNE", "TERMINE"]
-    }
-  },
-  "schemas": {},
-  "sequences": {},
-  "roles": {},
-  "policies": {},
-  "views": {},
-  "_meta": {
-    "columns": {},
-    "schemas": {},
-    "tables": {}
-  }
-}
