+++ api/drizzle/meta/_journal.json
@@ -5,15 +5,8 @@
     {
       "idx": 0,
       "version": "7",
-      "when": 1736432145404,
-      "tag": "0000_early_bushwacker",
-      "breakpoints": true
-    },
-    {
-      "idx": 2,
-      "version": "7",
-      "when": 1736845179626,
-      "tag": "0002_fluffy_power_man",
+      "when": 1736862406795,
+      "tag": "0000_tan_lady_ursula",
       "breakpoints": true
     }
   ]
