+++ api/src/projects/projects.controller.ts
@@ -1,16 +1,22 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Req } from "@nestjs/common";
-import { ProjectsService } from "./services/projects.service";
+import { Body, Controller, Get, Param, Patch, Post, Req } from "@nestjs/common";
 import { CreateOrUpdateProjectResponse, CreateProjectRequest } from "./dto/create-project.dto";
 import { UpdateProjectDto } from "./dto/update-project.dto";
 import { ProjectResponse } from "./dto/project.dto";
 import { ApiBearerAuth } from "@nestjs/swagger";
 import { ApiEndpointResponses } from "@/shared/decorator/api-response.decorator";
 import { Request } from "express";
+import { CreateProjectsService } from "./services/create-projects/create-projects.service";
+import { UpdateProjectsService } from "./services/update-projects/update-projects.service";
+import { GetProjectsService } from "@projects/services/get-projects/get-projects.service";
 
 @ApiBearerAuth()
 @Controller("projects")
 export class ProjectsController {
-  constructor(private readonly projectsService: ProjectsService) {}
+  constructor(
+    private readonly projectCreateService: CreateProjectsService,
+    private readonly projectFindService: GetProjectsService,
+    private readonly projectUpdateService: UpdateProjectsService,
+  ) {}
 
   @Post()
   @ApiEndpointResponses({
@@ -22,7 +28,7 @@ export class ProjectsController {
     @Req() _request: Request,
     @Body() createProjectDto: CreateProjectRequest,
   ): Promise<CreateOrUpdateProjectResponse> {
-    return this.projectsService.create(createProjectDto);
+    return this.projectCreateService.create(createProjectDto);
   }
 
   @Get()
@@ -32,13 +38,13 @@ export class ProjectsController {
     isArray: true,
   })
   findAll(): Promise<ProjectResponse[]> {
-    return this.projectsService.findAll();
+    return this.projectFindService.findAll();
   }
 
   @ApiEndpointResponses({ successStatus: 200, response: ProjectResponse })
   @Get(":id")
   findOne(@Param("id") id: string): Promise<ProjectResponse> {
-    return this.projectsService.findOne(id);
+    return this.projectFindService.findOne(id);
   }
 
   @Patch(":id")
@@ -52,12 +58,6 @@ export class ProjectsController {
     @Param("id") id: string,
     @Body() updateProjectDto: UpdateProjectDto,
   ): Promise<CreateOrUpdateProjectResponse> {
-    return this.projectsService.update(id, updateProjectDto);
-  }
-
-  //todo to implement
-  @Delete(":id")
-  remove(@Param("id") id: string): string {
-    return this.projectsService.remove(id);
+    return this.projectUpdateService.update(id, updateProjectDto);
   }
 }
