+++ api/src/projects/services/get-projects/get-projects.service.spec.ts
@@ -1,16 +1,18 @@
 // disabled to use expect any syntax
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { ProjectsService } from "./projects.service";
+
 import { TestDatabaseService } from "@test/helpers/test-database.service";
 import { teardownTestModule, testModule } from "@test/helpers/testModule";
-import { CreateProjectRequest } from "../dto/create-project.dto";
+import { CreateProjectRequest } from "../../dto/create-project.dto";
 import { TestingModule } from "@nestjs/testing";
 import { NotFoundException } from "@nestjs/common";
 import { getFutureDate } from "@test/helpers/getFutureDate";
-import { UpdateProjectDto } from "@projects/dto/update-project.dto";
+import { GetProjectsService } from "./get-projects.service";
+import { CreateProjectsService } from "../create-projects/create-projects.service";
 
-describe("ProjectsService", () => {
-  let service: ProjectsService;
+describe("ProjectFindService", () => {
+  let findService: GetProjectsService;
+  let createService: CreateProjectsService;
   let testDbService: TestDatabaseService;
   let module: TestingModule;
 
@@ -20,7 +22,8 @@ describe("ProjectsService", () => {
     const { module: internalModule, testDbService: tds } = await testModule();
     module = internalModule;
     testDbService = tds;
-    service = module.get<ProjectsService>(ProjectsService);
+    findService = module.get<GetProjectsService>(GetProjectsService);
+    createService = module.get<CreateProjectsService>(CreateProjectsService);
   });
 
   afterAll(async () => {
@@ -31,25 +34,6 @@ describe("ProjectsService", () => {
     await testDbService.cleanDatabase();
   });
 
-  describe("create", () => {
-    it("should create a new project", async () => {
-      const createDto: CreateProjectRequest = {
-        nom: "Test Project",
-        description: "Test Description",
-        budget: 100000,
-        forecastedStartDate: getFutureDate(),
-        status: "IDEE",
-        communeInseeCodes: mockedCommunes,
-        competencesAndSousCompetences: ["Santé", "Culture__Arts plastiques et photographie"],
-      };
-
-      const result = await service.create(createDto);
-
-      expect(result).toEqual({
-        id: expect.any(String),
-      });
-    });
-  });
   describe("findAll", () => {
     it("should return all projects", async () => {
       const futureDate = getFutureDate();
@@ -71,8 +55,8 @@ describe("ProjectsService", () => {
         communeInseeCodes: mockedCommunes,
       };
 
-      await service.create(createDto1);
-      await service.create(createDto2);
+      await createService.create(createDto1);
+      await createService.create(createDto2);
 
       const {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -85,7 +69,7 @@ describe("ProjectsService", () => {
         ...expectedFieldsProject2
       } = createDto2;
 
-      const result = await service.findAll();
+      const result = await findService.findAll();
 
       const expectedCommonFields = {
         id: expect.any(String),
@@ -134,8 +118,8 @@ describe("ProjectsService", () => {
         communeInseeCodes: mockedCommunes,
       };
 
-      const createdProject = await service.create(createDto);
-      const result = await service.findOne(createdProject.id);
+      const createdProject = await createService.create(createDto);
+      const result = await findService.findOne(createdProject.id);
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const { communeInseeCodes, competencesAndSousCompetences, ...expectedFields } = createDto;
       expect(result).toEqual({
@@ -160,89 +144,7 @@ describe("ProjectsService", () => {
 
     it("should throw NotFoundException when project not found", async () => {
       const nonExistentId = "00000000-0000-0000-0000-000000000000";
-      await expect(service.findOne(nonExistentId)).rejects.toThrow(NotFoundException);
-      await expect(service.findOne(nonExistentId)).rejects.toThrow(`Project with ID ${nonExistentId} not found`);
-    });
-  });
-
-  describe("update", () => {
-    let projectId: string;
-
-    beforeEach(async () => {
-      const createDto: CreateProjectRequest = {
-        nom: "Initial Project",
-        description: "Initial Description",
-        porteurReferentEmail: "initial@email.com",
-        budget: 100000,
-        forecastedStartDate: getFutureDate(),
-        status: "IDEE",
-        communeInseeCodes: mockedCommunes,
-      };
-
-      const result = await service.create(createDto);
-      projectId = result.id;
-    });
-
-    it("should update basic project fields", async () => {
-      const updateDto = {
-        nom: "Updated Project",
-        description: "Updated Description",
-        budget: 200000,
-      };
-
-      await service.update(projectId, updateDto);
-
-      const updatedProject = await service.findOne(projectId);
-
-      expect(updatedProject).toMatchObject({
-        ...updateDto,
-        id: projectId,
-        porteurReferentEmail: "initial@email.com",
-        communes: expect.arrayContaining(
-          mockedCommunes.map((code) => ({
-            inseeCode: code,
-          })),
-        ),
-      });
-    });
-
-    it("should only update communes when this is the only change", async () => {
-      const newCommunes = ["34567", "89012"];
-      const updateDto: UpdateProjectDto = {
-        communeInseeCodes: newCommunes,
-      };
-
-      await service.update(projectId, updateDto);
-      const updatedProject = await service.findOne(projectId);
-
-      expect(updatedProject.communes).toHaveLength(newCommunes.length);
-      expect(updatedProject.communes).toEqual(
-        expect.arrayContaining(
-          newCommunes.map((code) => ({
-            inseeCode: code,
-          })),
-        ),
-      );
-    });
-
-    it("should throw NotFoundException when project doesn't exist", async () => {
-      const nonExistentId = "00000000-0000-0000-0000-000000000000";
-      const updateDto = { nom: "Updated Name" };
-
-      await expect(service.update(nonExistentId, updateDto)).rejects.toThrow(NotFoundException);
-      await expect(service.update(nonExistentId, updateDto)).rejects.toThrow(
-        `Project with ID ${nonExistentId} not found`,
-      );
-    });
-
-    it("should validate forecasted start date", async () => {
-      const pastDate = new Date();
-      pastDate.setFullYear(pastDate.getFullYear() - 1);
-      const updateDto = {
-        forecastedStartDate: pastDate.toISOString().split("T")[0],
-      };
-
-      await expect(service.update(projectId, updateDto)).rejects.toThrow("Forecasted start date must be in the future");
+      await expect(findService.findOne(nonExistentId)).rejects.toThrow(NotFoundException);
     });
   });
 });
