+++ api/src/projects/services/update-projects/update-projects.service.ts
@@ -1,0 +1,66 @@
+import { DatabaseService } from "@database/database.service";
+import { projects } from "@database/schema";
+import { UpdateProjectDto } from "@projects/dto/update-project.dto";
+import { removeUndefined } from "@/shared/utils/remove-undefined";
+import { Injectable, NotFoundException, BadRequestException } from "@nestjs/common";
+import { eq } from "drizzle-orm";
+import { CommunesService } from "../communes/communes.service";
+import { CompetencesService } from "@projects/services/competences/competences.service";
+
+@Injectable()
+export class UpdateProjectsService {
+  constructor(
+    private dbService: DatabaseService,
+    private readonly communesService: CommunesService,
+    private readonly competencesService: CompetencesService,
+  ) {}
+
+  async update(id: string, updateProjectDto: UpdateProjectDto): Promise<{ id: string }> {
+    if (updateProjectDto.forecastedStartDate) {
+      this.validateDate(updateProjectDto.forecastedStartDate);
+    }
+    const { competencesAndSousCompetences, ...otherFields } = updateProjectDto;
+    const { competences, sousCompetences } = this.competencesService.splitCompetence(competencesAndSousCompetences);
+
+    return this.dbService.database.transaction(async (tx) => {
+      const [existingProject] = await tx
+        .select({
+          id: projects.id,
+          porteurReferentEmail: projects.porteurReferentEmail,
+        })
+        .from(projects)
+        .where(eq(projects.id, id))
+        .limit(1);
+
+      if (!existingProject) {
+        throw new NotFoundException(`Project with ID ${id} not found`);
+      }
+
+      const { communeInseeCodes, ...fieldsToUpdate } = removeUndefined({
+        ...otherFields,
+        competences,
+        sousCompetences,
+      });
+
+      if (communeInseeCodes) {
+        await this.communesService.createOrUpdate(tx, id, communeInseeCodes);
+      }
+
+      if (Object.keys(fieldsToUpdate).length > 0) {
+        await tx.update(projects).set(fieldsToUpdate).where(eq(projects.id, id));
+      }
+
+      return { id: existingProject.id };
+    });
+  }
+
+  private validateDate(dateStr: string): void {
+    const inputDate = new Date(dateStr);
+    const today = new Date();
+    today.setHours(0, 0, 0, 0);
+
+    if (inputDate < today) {
+      throw new BadRequestException("Forecasted start date must be in the future");
+    }
+  }
+}
